# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "System Main Data",
        "contact": {},
        "license": {}
    },
    "paths": {
        "/blocks": {
            "get": {
                "description": "ListBlocks blocks by list parameters",
                "tags": [
                    "Block"
                ],
                "summary": "ListBlocks Blocks",
                "operationId": "list-blocks",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "schema": {
                            "$ref": "#/definitions/viewmodels.ListBlocks"
                        }
                    },
                    "500": {
                        "description": "Internal error.",
                        "schema": {
                            "$ref": "#/definitions/viewmodels.Error"
                        }
                    }
                }
            }
        },
        "/blocks/{block_num}": {
            "get": {
                "description": "Get block by block number",
                "tags": [
                    "Block"
                ],
                "summary": "Get Block",
                "operationId": "get-block",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Block Number",
                        "name": "block_num",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "schema": {
                            "$ref": "#/definitions/viewmodels.GetSingleBlock"
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "$ref": "#/definitions/viewmodels.Error"
                        }
                    },
                    "500": {
                        "description": "Internal error.",
                        "schema": {
                            "$ref": "#/definitions/viewmodels.Error"
                        }
                    }
                }
            }
        },
        "/transaction/{tx_hash}": {
            "get": {
                "description": "GetBlock block by block number",
                "tags": [
                    "Transaction"
                ],
                "summary": "Get Transaction",
                "operationId": "get-transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction hash",
                        "name": "tx_hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "schema": {
                            "$ref": "#/definitions/viewmodels.GetTransaction"
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "$ref": "#/definitions/viewmodels.Error"
                        }
                    },
                    "500": {
                        "description": "Internal error.",
                        "schema": {
                            "$ref": "#/definitions/viewmodels.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "inputmodels.ListBlocksParam": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Limit",
                    "type": "integer"
                }
            }
        },
        "viewmodels.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "viewmodels.GetSingleBlock": {
            "type": "object",
            "properties": {
                "block_hash": {
                    "type": "string"
                },
                "block_num": {
                    "type": "integer"
                },
                "block_time": {
                    "type": "string"
                },
                "parent_hash": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "viewmodels.GetTransaction": {
            "type": "object",
            "properties": {
                "block_hash": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodels.TransactionLog"
                    }
                },
                "nonce": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                },
                "tx_hash": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "viewmodels.ListBlockItem": {
            "type": "object",
            "properties": {
                "block_hash": {
                    "type": "string"
                },
                "block_num": {
                    "type": "integer"
                },
                "block_time": {
                    "type": "string"
                },
                "parent_hash": {
                    "type": "string"
                }
            }
        },
        "viewmodels.ListBlocks": {
            "type": "object",
            "properties": {
                "blocks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodels.ListBlockItem"
                    }
                }
            }
        },
        "viewmodels.TransactionLog": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "tx_hash": {
                    "type": "string"
                }
            }
        }
    }
}